refactor errors to errors.js file
